+++ extensions/parquet/table/src/test/java/io/deephaven/parquet/table/S3ParquetLocalStackTest.java
@@ -4,31 +4,26 @@
 package io.deephaven.parquet.table;
 
 import io.deephaven.extensions.s3.S3Instructions.Builder;
-import io.deephaven.extensions.s3.testlib.SingletonContainers.MinIO;
+import io.deephaven.extensions.s3.testlib.SingletonContainers.LocalStack;
 import io.deephaven.extensions.s3.testlib.SingletonContainers;
-import io.deephaven.stats.util.OSUtil;
 import org.junit.BeforeClass;
 import software.amazon.awssdk.services.s3.S3AsyncClient;
 
-import static org.junit.Assert.assertFalse;
-
-public class ParquetS3SimpleMinIOTest extends ParquetS3SimpleTestBase {
+public class S3ParquetLocalStackTest extends S3ParquetTestBase {
 
     @BeforeClass
     public static void initContainer() {
-        // TODO(deephaven-core#5116): MinIO testcontainers does not work on OS X
-        assertFalse("OSUtil.runningMacOS()", OSUtil.runningMacOS());
         // ensure container is started so container startup time isn't associated with a specific test
-        MinIO.init();
+        LocalStack.init();
     }
 
     @Override
-    public Builder s3Instructions(final Builder builder) {
-        return MinIO.s3Instructions(builder);
+    public Builder s3Instructions(Builder builder) {
+        return LocalStack.s3Instructions(builder);
     }
 
     @Override
     public S3AsyncClient s3AsyncClient() {
-        return SingletonContainers.MinIO.s3AsyncClient();
+        return SingletonContainers.LocalStack.s3AsyncClient();
     }
 }
